requirements/local.txt
# Include base requirements
-r base.txt

django_debug_toolbar==1.2.1
sqlparse==0.1.1
